+++ src/Db4Net/Commands/UpdateCommandBuilder.cs
@@ -51,7 +51,7 @@ public sealed class UpdateCommandBuilder<T> : CommandBuilderBase
 
     internal UpdateCommandBuilder<T> SetEntityValues(T entity)
     {
-        ArgumentNullException.ThrowIfNull(entity);
+        ThrowHelper.ThrowIfNull(entity);
 
         foreach (var column in ModelMetadata<T>.NonKeyColumns)
         {
@@ -131,7 +131,7 @@ public sealed class UpdateCommandBuilder<T> : CommandBuilderBase
     /// <returns>The current command builder.</returns>
     public UpdateCommandBuilder<T> WhereKey(T entity)
     {
-        ArgumentNullException.ThrowIfNull(entity);
+        ThrowHelper.ThrowIfNull(entity);
 
         foreach (var column in ModelMetadata<T>.RequireKeyColumns())
         {
@@ -229,7 +229,7 @@ public sealed class UpdateCommandBuilder<T> : CommandBuilderBase
 
     private void AddGroup(FilterBooleanOperator booleanOperator, Action<FilterGroupBuilder<T>> configure)
     {
-        ArgumentNullException.ThrowIfNull(configure);
+        ThrowHelper.ThrowIfNull(configure);
 
         var group = new FilterGroupBuilder<T>();
         configure(group);