+++ src/Db4Net/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using Db4Net.Internal;